remove unnecessary diffs
